+++ src/L10NSharp/TranslationMemory.cs
@@ -1,11 +1,11 @@
-// Copyright (c) 2019 SIL International
+// Copyright (c) 2022 SIL International
 // This software is licensed under the MIT License (http://opensource.org/licenses/MIT)
 
 namespace L10NSharp
 {
 	public enum TranslationMemory
 	{
-		Tmx,
+		//Tmx, - no longer supported
 		XLiff
 	}
 }